fix shitcode
