--- Content.Server/Atmos/EntitySystems/GasTankSystem.cs
@@ -1,32 +1,0 @@
-using Content.Server.Atmos.Components;
-using JetBrains.Annotations;
-using Robust.Shared.GameObjects;
-
-namespace Content.Server.Atmos.EntitySystems
-{
-    [UsedImplicitly]
-    public class GasTankSystem : EntitySystem
-    {
-        private float _timer = 0f;
-        private const float Interval = 0.5f;
-
-        public override void Update(float frameTime)
-        {
-            base.Update(frameTime);
-
-            _timer += frameTime;
-
-            if (_timer < Interval) return;
-            _timer = 0f;
-
-            var atmosphereSystem = Get<AtmosphereSystem>();
-
-            foreach (var gasTank in EntityManager.ComponentManager.EntityQuery<GasTankComponent>(true))
-            {
-                atmosphereSystem.React(gasTank.Air, gasTank);
-                gasTank.CheckStatus();
-                gasTank.UpdateUserInterface();
-            }
-        }
-    }
-}
